requirements.txt
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
asgiref==3.8.1
asn1crypto==1.5.1
beautifulsoup4==4.12.3
bcrypt==3.2.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
Django==3.2.12
django-cors-headers==4.3.1
django-rest-swagger==2.2.0
djangorestframework==3.15.1
drf-yasg==1.21.7
ecdsa==0.19.0
fastapi==0.110.2
greenlet==3.0.3
google-search-results==2.4.2
gunicorn==22.0.0
h11==0.14.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.0
passlib==1.7.4
pg8000==1.31.2
pipdeptree==2.18.1
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.0
pydantic[email]==2.7.0
pydantic_core==2.18.1
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
schedule==1.2.2
scramp==1.4.5
setuptools==68.2.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.5.0
starlette==0.37.2
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
wheel==0.41.2
python-dotenv==1.0.1
groq==0.13.0